# chapter-05/project/app/Http/Controllers/API/BuilderController.php
<?php

namespace App\Http\Controllers\API;

use Illuminate\Http\Request;
use App\Http\Controllers\Controller;
use App\Builder;

class BuilderController extends Controller
{
    /**
     * Display a listing of the resource.
     *
     * @return \Illuminate\Http\Response
     * 
     *  @SWG\Get(
     *      path="/api/builders",
     *      tags={"Builders"},
     *      summary="List Builders",
     *      @SWG\Response(
     *          response=200,
     *          description="Success: List all Builders",
     *          @SWG\Schema(ref="#/definitions/Builder")
     *      ),
     *      @SWG\Response(
     *          response="404",
     *          description="Not Found",
     *      )
     * )
     */
    public function index()
    {
        $listBuilder = Builder::all();
        return $listBuilder;
    }

    /**
     * Store a newly created resource in storage.
     *
     * @param  \Illuminate\Http\Request  $request
     * @return \Illuminate\Http\Response
     * 
     *  @SWG\Post(
     *      path="/api/builders",
     *      tags={"Builders"},
     *      summary="List Builder",
     *      @SWG\Parameter(
     *          name="body",
     *          in="body",
     *          required=true,
     *          @SWG\Schema(ref="#/definitions/Builder"),
     *          description="Json format",
     *      ),
     *      @SWG\Response(
     *          response=201,
     *          description="Success: A Newly Created Builder",
     *          @SWG\Schema(ref="#/definitions/Builder")
     *      ),
     *      @SWG\Response(
     *          response="422",
     *          description="Missing mandatory field",
     *      ),
     *      @SWG\Response(
     *          response="404",
     *          description="Not Found",
     *      ),
     *      @SWG\Response(
     *          response="405",
     *          description="Invalid HTTP Method",
     *      )
     * )
     */
    public function store(Request $request)
    {
        $createBuilder = Builder::create($request->all());
        return $createBuilder;
    }

    /**
     * Display the specified resource.
     *
     * @param  int  $id
     * @return \Illuminate\Http\Response
     * 
     *  @SWG\Get(
     *      path="/api/builders/{id}",
     *      tags={"Builders"},
     *      summary="Get Builder by Id",
     *      @SWG\Parameter(
     *          name="id",
     *          in="path",
     *          required=true,
     *          type="integer",
     *          description="Display thhe specified Builder by id.",
     *      ),
     *      @SWG\Response(
     *          response=200,
     *          description="Success: Return the Builder",
     *          @SWG\Schema(ref="#/definitions/Builder")
     *      ),
     *      @SWG\Response(
     *          response="404",
     *          description="Not Found",
     *      ),
     *      @SWG\Response(
     *          response="405",
     *          description="Invalid HTTP Method",
     *      ),
     * )
     */
    public function show($id)
    {
        $showBuilderById = Builder::with('Bike')->findOrFail($id);
        return $showBuilderById;
    }

    /**
     * Update the specified resource in storage.
     *
     * @param  \Illuminate\Http\Request  $request
     * @param  int  $id
     * @return \Illuminate\Http\Response
     * 
     *  @SWG\Put(
     *      path="/api/builders/{id}",
     *      tags={"Builders"},
     *      summary="Update Builder",
     *      @SWG\Parameter(
     *          name="id",
     *          in="path",
     *          required=true,
     *          type="integer",
     *          description="Update the specified Builder by id.",
     *      ),
     *      @SWG\Parameter(
     *          name="body",
     *          in="body",
     *          required=true,
     *          @SWG\Schema(ref="#/definitions/Builder"),
     *          description="Json format",
     *      ),
     *      @SWG\Response(
     *          response=200,
     *          description="Success: Return the Builder updated",
     *          @SWG\Schema(ref="#/definitions/Builder")
     *      ),
     *      @SWG\Response(
     *          response="422",
     *          description="Missing mandatory field",
     *      ),
     *      @SWG\Response(
     *          response="404",
     *          description="Not Found",
     *      ),
     *      @SWG\Response(
     *          response="405",
     *          description="Invalid HTTP Method",
     *      ),
     * ) 
     */
    public function update(Request $request, $id)
    {
        $updateBuilderById = Builder::findOrFail($id);
        $updateBuilderById->update($request->all());
        return $updateBuilderById;
    }

    /**
     * Remove the specified resource from storage.
     *
     * @param  int  $id
     * @return \Illuminate\Http\Response
     * 
     *  @SWG\Delete(
     *      path="/api/builders/{id}",
     *      tags={"Builders"},
     *      summary="Delete Builder",
     *      description="Delete the specified Builder by id",
     *      @SWG\Parameter(
     *          name="id",
     *          in="path",
     *          required=true,
     *          type="integer",
     *          description="Builder id to delete",
     *          format="int64"
     *      ),
     *      @SWG\Response(
     *          response="404",
     *          description="Not Found",
     *      ),
     *      @SWG\Response(
     *          response="405",
     *          description="Invalid HTTP Method",
     *      ),
     *      @SWG\Response(
     *          response="204",
     *          description="Success: successful deleted",
     *      ),
     *  )
     */
    public function destroy($id)
    {
        $deleteBikeById = Bike::find($id)->delete();
        return response()->json([], 204);
    }
}